+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -1,11 +1,9 @@
-package chess.domain.position;
+package chess.domain;
 
 import static chess.constants.TestConstants.PARAMETERIZED_TEST_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.ChessBoard;
-import chess.domain.GameCommand;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
