+++ src/test/java/chess/domain/position/DirectionTest.java
@@ -1,12 +1,10 @@
-package chess.domain;
+package chess.domain.position;
 
 import static chess.constants.TestConstants.PARAMETERIZED_TEST_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import chess.domain.piece.Color;
-import chess.domain.position.Direction;
-import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
